\name{dprop-package}
\alias{dprop-package}
\docType{package}
\title{
Computation of Some Important Distributional Properties
}

\description{
Generally, most of the packages specify the probability density function, cumulative distribution function, quantile function, and random numbers generation of the probability distributions. The present package allows to compute some important distributional properties, including the first four ordinary and central moments, Pearson's coefficient of skewness and kurtosis, the mean and variance, coefficient of variation, median, and quartile deviation at some parametric values of several well-known and extensively used probability distributions.
}

\details{
\tabular{ll}{
Package: \tab dprop\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2023-06-28\cr
License: \tab GPL-2\cr
}
}

\author{
 Christophe Chesneau <christophe.chesneau@unicaen.fr>, Muhammad Imran <imranshakoor84@yahoo.com>, M.H Tahir <mht@iub.edu.pk> and Farrukh Jamal <farrukh.jamal@iub.edu.pk>.
}

\section{Maintainers}{
Muhammad Imran <imranshakoor84@yahoo.com>
}





\keyword{Distribution theory}
\keyword{Central moments}
\keyword{Ordinary moments}
\keyword{Skewness}
\keyword{Kurtosis}
\keyword{Mean and variance}
\keyword{Quartile deviation}
\keyword{Median}
\keyword{Coefficient of variation}



